Use indexed array for included clients
The change is cosmetic, but it allows administrators to easily add extra
clients in the configuration file, with:

	include 'clients.php' + [[…], […], …]

This works because the `+` operator merges arrays but removes duplicate
array keys.  Since the included clients have their client_id as key,
it will never conflict with a client defined by the administrator,
even if the administrator does not use keys and thus the keys are set
automatically to incrementing integers.